% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.array.R
\name{as.array.ecmeta}
\alias{as.array.ecmeta}
\alias{as.array.ecmeta_jags}
\alias{as.array.ecmeta_jags_prediction}
\alias{as.array.ecmeta_ml_prediction}
\title{Convert object to array}
\usage{
\method{as.array}{ecmeta_jags}(x, ...)

\method{as.array}{ecmeta_jags_prediction}(x, ...)

\method{as.array}{ecmeta_ml_prediction}(x, ...)
}
\arguments{
\item{x}{An object of the appropriate class.}

\item{...}{Additional arguments to pass to \code{\link[coda:mcmc.convert]{coda::as.array.mcmc.list()}} other
than \code{drop}, which is set to \code{FALSE}.}
}
\value{
A 3-D array in the format described in \code{\link[bayesplot:MCMC-overview]{bayesplot::MCMC}}.
}
\description{
This is a helper function to convert a object to a 3-D
array as required by \code{\link{bayesplot}} in order to diagnosis the
Markov Chain Monte Carlo (MCMC) sampling. See \code{\link[bayesplot:MCMC-overview]{bayesplot::MCMC}} for more
information about the format of the array.
}
