% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-objects.R
\name{as_loghr_data}
\alias{as_loghr_data}
\alias{as_loghr_data.data.frame}
\title{Convert to \code{loghr_data}}
\usage{
as_loghr_data(x, ...)

\method{as_loghr_data}{data.frame}(x, estimate = "estimate", standard_error = "standard_error", ...)
}
\arguments{
\item{x}{An \code{R} object.}

\item{estimate}{Name of the column containing the point estimate of
the log hazard ratio.}

\item{standard_error}{Name of the column containing the standard error of
the estimate of the log hazard ratio.}
}
\description{
Methods to convert an object to a \code{\link{loghr_data}} object.
}
