% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecmeta.R
\name{ecmeta}
\alias{ecmeta}
\title{Meta-analysis of external control studies}
\usage{
ecmeta(data, method = c("jags", "ml"), ...)
}
\arguments{
\item{data}{A \code{\link{loghr_data}} object that stores estimates of the log hazard ratios
of the internal control arms relative to the external control arms.}

\item{method}{Method/software used to estimate the meta-analytic model. \code{jags}
uses Markov Chain Monte Carlo to sample from the posterior distribution of
the parameters while \code{ml} using maximum likelihood to fit the model.}

\item{...}{Additional arguments to pass to the underlying meta-analytic estimation
function. If \code{method == "jags"}, these are arguments to pass to \code{\link[=ecmeta_jags]{ecmeta_jags()}}.
Unused if \verb{method == "ml}.}
}
\value{
Returns an object with class based on \code{method}. If \code{method == "jags"},
then an \code{\link{ecmeta_jags}} object is returned; if \verb{method == "ml}, then
an \code{\link{ecmeta_ml}} object is returned.
}
\description{
Perform a meta-analysis of studies in which external control arms were
compared to internal control arms. For more information on the methodology
please see \code{vignette("methodology")}.
}
\seealso{
See \code{vignette("guide")} for an example.
}
