% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecmeta_jags.R
\name{ecmeta_jags}
\alias{ecmeta_jags}
\alias{print.ecmeta_jags}
\title{Meta-analysis of external control studies using JAGS}
\usage{
ecmeta_jags(
  data,
  n_chains = 1,
  n_iter = 1000,
  thin = 1,
  n_burnin = 0,
  n_adapt = 1000,
  quiet = FALSE,
  prior_mu = normal(0, 100),
  prior_scale = invgamma(0.001, 0.001)
)
}
\arguments{
\item{data}{A \code{\link{loghr_data}} object that stores estimates of the log hazard ratios
of the internal control arms relative to the external control arms.}

\item{n_chains}{The number of parallel chains for the model. Passed to
\code{\link[rjags:jags.model]{rjags::jags.model()}}.}

\item{n_iter}{Number of iterations to monitor. Passed to \code{\link[rjags:coda.samples]{rjags::coda.samples()}}.}

\item{thin}{Thinning interval for monitors. Passed to \code{\link[rjags:coda.samples]{rjags::coda.samples()}}.}

\item{n_burnin}{Number of}

\item{n_adapt}{The number of iterations for adaptation. Passed to \code{\link[rjags:jags.model]{rjags::jags.model()}}.}

\item{quiet}{If \code{TRUE} then messages generated during compilation will be suppressed,
as well as the progress bar during adaptation. Passed to \code{\link[rjags:jags.model]{rjags::jags.model()}}.}

\item{prior_mu}{Prior for \code{mu}. Must be specified as a normal distribution with
\code{\link[=normal]{normal()}}.}

\item{prior_scale}{Prior for the scale parameter, which is either \code{sigma} or
\code{sigma^2} depending on the choice of prior. A half \code{\link[=student_t]{student_t()}} distribution
or uniform prior can be placed on \code{sigma} while an \code{\link[=invgamma]{invgamma()}} prior can be
used for \code{sigma^2}.}
}
\value{
An \code{ecmeta_jags} object, which is a list containing the following elements:
\describe{
\item{mcmc}{A \code{\link[coda:mcmc.list]{coda::mcmc.list}} object containing samples of the model parameters.}
\item{data}{The \code{data} object.}
}
}
\description{
Perform a meta-analysis of studies in which external control arms were
compared to internal control arms using \code{\link{rjags}}. You can view the
\code{\link[=ecmeta]{ecmeta()}} documentation for more general details.
}
