% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecmeta_ml.R
\name{ecmeta_ml}
\alias{ecmeta_ml}
\alias{print.ecmeta_ml}
\title{Meta-analysis of external control studies using maximum likelihood}
\usage{
ecmeta_ml(data, hessian = TRUE)
}
\arguments{
\item{data}{A \code{\link{loghr_data}} object that stores estimates of the log hazard ratios
of the internal control arms relative to the external control arms.}

\item{hessian}{Logical passed to \code{\link[stats:optim]{stats::optim()}} indicating whether a Hessian
matrix should be computed. This is required for computation of standard errors
for \code{mu} and \code{sigma}, but not needed to simulate predicted log hazard ratios
with \code{\link[=predict.ecmeta_ml]{predict.ecmeta_ml()}}.}
}
\value{
An \code{ecmeta_ml} object which is a list containing the following elements:
\describe{
\item{estimates}{Point estimates of model parameters \code{mu} and \code{logsigma} (the
log of \code{sigma}).}
\item{vcov}{The variance covariance matrix of the model parameters. Only included
if \code{hessian = TRUE}.}
\item{data}{The \code{data} object.}
}
}
\description{
Perform a meta-analysis of studies in which external control arms were
compared to internal control arms by optimizing the log-likelihood function.
You can view the \code{\link[=ecmeta]{ecmeta()}} documentation for more general details.
}
\seealso{
\code{\link[=ecmeta]{ecmeta()}}
}
