% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-objects.R
\name{loghr_data}
\alias{loghr_data}
\title{Create a log hazard ratio data object}
\usage{
loghr_data(estimate, standard_error)
}
\arguments{
\item{estimate}{The point estimate of the log hazard ratio.}

\item{standard_error}{The standard error of the estimate of the
log hazard ratio.}
}
\value{
An object of class \code{loghr_data}, which inherits from \code{data.frame} and
contains one column named \code{estimate} and a second column named \code{standard_error}.
}
\description{
Create a data object that stores estimates of log hazard ratios from a survival
model or covert an existing object with \code{as_loghr_data()}.
}
\seealso{
\code{\link{loghr_data}} objects can also be created from existing \code{R} objects
with \code{\link[=as_loghr_data]{as_loghr_data()}}.
}
