% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ecmeta}
\alias{predict.ecmeta}
\alias{print.ecmeta_jags_prediction}
\alias{predict.ecmeta_jags}
\alias{predict.ecmeta_ml}
\alias{print.ecmeta_ml_prediction}
\title{Predict log hazard ratio}
\usage{
\method{predict}{ecmeta_jags}(
  object,
  newdata,
  n_burnin = 0,
  thin = 1,
  n_adapt = 1000,
  quiet = FALSE,
  ...
)

\method{predict}{ecmeta_ml}(object, newdata, n_sims = 1000, ...)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{newdata}{A \code{loghr_data} object that stores estimate of the log hazard
ratio for a comparison of the treatment arm to the external control arm in
the new single-arm study.}

\item{n_burnin}{Number of}

\item{thin}{Thinning interval for monitors. Passed to \code{\link[rjags:coda.samples]{rjags::coda.samples()}}.}

\item{n_adapt}{The number of iterations for adaptation. Passed to \code{\link[rjags:jags.model]{rjags::jags.model()}}.}

\item{quiet}{If \code{TRUE} then messages generated during compilation will be suppressed,
as well as the progress bar during adaptation. Passed to \code{\link[rjags:jags.model]{rjags::jags.model()}}.}

\item{...}{Currently unused.}

\item{n_sims}{Number of simulations to use. Only relevant when using a
maximum likelihood based approach.}
}
\value{
A list that may contain the following elements:
\describe{
\item{loghr}{A matrix with three columns containing draws of the true log
hazard ratios. The columns are: \code{trt_ic} (a comparison of the treatment
to the internal control), \code{trt_ec} (a comparison of the treatment to the
external control) and \code{ic_ec} (a comparison of the internal control to
the external control).}
\item{mcmc}{When using a Bayesian approach, a \code{\link[coda:mcmc.list]{coda::mcmc.list}} object is
also included that contains posterior samples of the true log hazard ratio
comparing the treatment to the external control. This is not included when
using a maximum likelihood approach.}
}
}
\description{
Predict the true log hazard ratio in a new single-arm study by adjusting
for additional bias and variability caused by the non-randomized design. This
function first draws from the distribution of the true log hazard
ratio comparing the treatment to the external control (\code{trt_ec}). It then
draws the true log hazard ratio comparing the internal control to the external
control (\code{ic_ec}) from its predictive distribution. Lastly, \code{trt_ec} and
\code{ic_ec} are combined to produce simulated draws of the true log hazard ratio
in a comparison of the treatment to the internal control (\code{trt_ic}).
}
\details{
The implementation differs slightly between the Bayesian and maximum likelihood
approaches. First, when using a Bayesian approach, MCMC is use to sample
\code{trt_ec} whereas \code{trt_ec} is sampled from a normal distribution based on
the log hazard ratio estimates and standard errors in the new study. In
practice, these approaches will produce very similar results.

Second, when using the Bayesian approach, the posterior predictive
distribution of \code{ic_ec} is drawn using the posterior samples of
\code{mu} and \code{sigma} stored in \code{object}. In the maximum likelihood approach,
\code{ic_ec} is simulated by using the point estimates of \code{mu} and \code{sigma} and
drawing from a t distribution, which is the predictive distribution for a
future observation.
}
\seealso{
See \code{vignette("methodology")} for a description of the method.
}
