% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors}
\alias{priors}
\alias{normal}
\alias{invgamma}
\alias{student_t}
\alias{uniform}
\title{Prior distributions}
\usage{
normal(mean = 0, sd = 10)

invgamma(shape = 0.001, rate = 0.001)

student_t(location = 0, scale = 25, df = 1)

uniform(min = 0, max = 1)
}
\arguments{
\item{mean, location}{Prior for the location parameter. This typically the mean,
but for a Cauchy distribution (which is a \code{student_t} with \code{df = 1}),
it is the median and the mean does not exist.}

\item{sd, scale}{Prior for the scale parameter, which determines the
dispersion of the distribution.}

\item{shape, rate}{Shape and rate parameters for the inverse gamma distribution.}

\item{min, max}{Lower and upper limits of the distribution. Must be finite.}
}
\value{
A named list to be used inside \code{ecmeta()} and \code{\link{predict.ecmeta}}
functions.
}
\description{
A collection of functions used to specify priors to pass to Bayesian models.
}
