% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat.R
\name{rhat_jags}
\alias{rhat_jags}
\title{R-hat convergence diagnostic}
\usage{
rhat_jags(x)
}
\arguments{
\item{An}{\code{\link[coda:mcmc.list]{coda::mcmc.list}} object with more than one chain as required
by \code{\link[coda:gelman.diag]{coda::gelman.diag()}}.}
}
\value{
A vector containing R-hat for each parameter.
}
\description{
This is a generic method for computing the R-hat convergence diagnostic
from multiple chains using \code{\link[coda:gelman.diag]{coda::gelman.diag()}}. Values substantially above 1
indicate a lack of convergence.
}
