% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ecmeta.R
\name{summary.ecmeta_jags}
\alias{summary.ecmeta_jags}
\title{Summary of external control meta-analysis using JAGS}
\usage{
\method{summary}{ecmeta_jags}(object, quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975), ...)
}
\arguments{
\item{object}{An \code{\link[=ecmeta_jags]{ecmeta_jags}} object.}

\item{quantiles}{A vector of quantiles to evaluate for each variable.}

\item{...}{Additional arguments to pass to \code{\link[coda:summary.mcmc]{coda::summary.mcmc.list()}}.}
}
\value{
A \code{data.frame} with one row for each parameter (\code{mu} and \code{sigma}).
The columns contain the same summary statistics and quantiles as in
\code{\link[coda:summary.mcmc]{coda::summary.mcmc.list()}}. Statistics reported are the mean (\code{mean}),
standard deviation (\code{sd}), naive (ignoring autocorrelation of the chain)
standard error of the mean (\code{se_mean}), and time-series standard error
based on an estimate of the spectral density at 0 (\code{se_ts}). Additional
columns contain quantiles of the distribution as specified by \code{quantiles}.
Finally, if multiple chains were used, the column \code{rhat} is included, which
is the R-hat convergence diagnostic (values substantially above 1 indicate
a lack of convergence).
}
\description{
Summary statistics for a meta-analytic model fit using \code{\link[=ecmeta]{ecmeta()}} with
\code{method == "jags"}.
}
