% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ecmeta.R
\name{summary.ecmeta_ml}
\alias{summary.ecmeta_ml}
\title{Summary of external control meta-analysis using maximum likelihood}
\usage{
\method{summary}{ecmeta_ml}(object, level = 0.95, ...)
}
\arguments{
\item{object}{An \code{\link[=ecmeta_ml]{ecmeta_ml}} object.}

\item{level}{The confidence level. Default is \code{0.95} for a 95 percent
confidence interval.}

\item{...}{Currently unused.}
}
\value{
A \code{data.frame} with one row for each parameter (\code{mu} and \code{sigma}).
There is always one column named \code{estimate} containing the point estimates.
In addition, if a variance-covariance matrix was estimated, then there are
additional columns for the standard errors as well as the lower and upper
confidence limits.
}
\description{
Summary statistics for a meta-analytic model fit using \code{\link[=ecmeta]{ecmeta()}} with
\code{method == "ml"}.
}
\note{
\code{object} contains estimates of \code{logsigma} rather than \code{sigma}. Point
estimates and confidence intervals for \code{sigma} are computed by simply
exponentiating \code{logsigma}; the standard error for \code{sigma} is estimated
using the delta method.
}
