% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ecmeta_prediction.R
\name{summary.ecmeta_prediction}
\alias{summary.ecmeta_prediction}
\alias{summary.ecmeta_jags_prediction}
\alias{summary.ecmeta_ml_prediction}
\title{Summary of predicted log hazard ratios}
\usage{
\method{summary}{ecmeta_jags_prediction}(
  object,
  exponentiate = FALSE,
  quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  ...
)

\method{summary}{ecmeta_ml_prediction}(
  object,
  exponentiate = FALSE,
  quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  ...
)
}
\arguments{
\item{object}{A object of the appropriate class.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
the log HR. Defaults to \code{FALSE}; if \code{TRUE}, then HRs
are summarized.}

\item{...}{Additional arguments to pass to \code{\link[coda:summary.mcmc]{coda::summary.mcmc.list()}}.}
}
\value{
A \code{data.frame} with one row for each log HR comparison:
treatment vs. internal control (\code{loghr_trt_ic}),
treatment vs. external control (\code{loghr_trt_ec}), and
internal control vs. external control (\code{loghr_ic_ec}).

The columns contain summary statistics and quantiles.
Statistics reported are the mean (\code{mean}), standard deviation (\code{sd}),
and naive (ignoring autocorrelation across iterations) standard error
of the mean. Additional columns contain quantiles of the distribution as
specified by \code{quantiles}.

If a Bayesian model was used to draw from the posterior distribution of
\code{loghr_trt_ec} and multiple chains were used, then the column
\code{rhat} is included as well. This is the R-hat convergence diagnostic with
values substantially above 1 indicating a lack of convergence. Note that
\code{rhat} is only provided for \code{loghr_trt_ec}. Diagnostics for \code{loghr_ic_ec} should
be based on \code{mu} and \code{sigma} from \code{\link[=ecmeta]{ecmeta()}} and \code{loghr_trt_ic} is
deterministically related to \code{loghr_trt_ec} and \code{loghr_ic_ec}.
}
\description{
Summary statistics for true log hazard ratios (HRs) predicted using
\code{ecmeta} \link[=predict.ecmeta]{prediction functions}.
}
